+++ src/main/java/org/yinwang/pysonar/demos/Demo.java
@@ -376,9 +376,9 @@ public class Demo {
 
 
     private static void usage() {
-        $.msg("Usage:  java -jar pysonar-3.3.2.jar <file-or-dir> <output-dir>");
+        $.msg("Usage:  java -jar pysonar-3.3.3.jar <file-or-dir> <output-dir>");
         $.msg("Example that generates an index for a Python 3 standard library:");
-        $.msg(" java -jar pysonar-3.3.2.jar /usr/lib/python3 ./html");
+        $.msg(" java -jar pysonar-3.3.3.jar /usr/lib/python3 ./html");
         System.exit(0);
     }